requirements.txt
instaloader==4.7.1
facebook-sdk==3.1.0
Flask==2.0.1
python-dotenv=0.17.1